main.py
from typing import TypedDict, Annotated
from dotenv import load_dotenv
load_dotenv()

from langchain_core.messages import BaseMessage, HumanMessage, AIMessage
from langgraph.graph import END, StateGraph
from langgraph.graph.message import add_messages

from chains import generate_chain, reflect_chain

class MessageGraph(TypedDict):
    messages: Annotated[list[BaseMessage], add_messages]

REFLECT = "reflect"
GENERATE = "generate"

def generation_node(state):
    return {"messages": [generate_chain.invoke({"messages": state["messages"]})]}

# def reflection_node(state):
#     res = reflect_chain.invoke({"messages": state["messages"]})
#     return {"messages": [HumanMessage(content=res.content)]}

def reflection_node(state):
    messages = state["messages"]

    # Convert last AI message to HumanMessage (Anthropic requirement)
    if isinstance(messages[-1], AIMessage):
        messages = messages[:-1] + [HumanMessage(content=messages[-1].content)]

    res = reflect_chain.invoke({"messages": messages})

    return {
        "messages": [HumanMessage(content=res.content)]
    }

builder = StateGraph(state_schema=MessageGraph)
builder.add_node(GENERATE, generation_node)
builder.add_node(REFLECT, reflection_node)
builder.set_entry_point(GENERATE)

def should_continue(state: MessageGraph):
    if len(state["messages"]) >= 6:
        return END
    return REFLECT

builder.add_conditional_edges(GENERATE, should_continue, path_map={END:END, REFLECT:REFLECT})
builder.add_edge(REFLECT, GENERATE)

graph = builder.compile()
print(graph.get_graph().draw_mermaid())

if __name__ == '__main__':
    print("Hello Langraph")
    inputs = {
    "messages": [
        HumanMessage(content="""Make this tweet better:
@LangChainAI

- newly Tool Calling feature is seriously underrated.

After a long wait, it's here- making the implementation of agents across different models with function calling.

Made a video covering their newest blog post
""")
    ]
}

response = graph.invoke(inputs)
