pyproject.toml
[project]
name = "reflection-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Poorvi Yadav",email = "poorvi@aurigaresearch.com"}
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "python-dotenv (>=1.2.2,<2.0.0)",
    "langchain (>=1.2.12,<2.0.0)",
    "langgraph (>=1.1.2,<2.0.0)",
    "langchain-anthropic (>=1.3.4,<2.0.0)",
    "langsmith (>=0.7.17,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
